feat: rename fatpack command to tar-fat32
feat: rename helm command to helm-local

chore: apply pre-commit

chore: apply pre-commit
