+++ README.md
@@ -71,7 +71,7 @@ pytest
 각 명령어에 대한 자세한 문서는 `docs/` 디렉토리에서 확인할 수 있습니다:
 - [Helm Local 명령어](docs/helm-local.md)
 - [Docker Tar 명령어](docs/docker_tar.md)
-- [FatPack 명령어](docs/fatpack.md)
+- [Tar-Fat32 명령어](docs/tar-fat32.md)
 - [S3 공유 명령어](docs/s3-share.md)
 - [PyPI 등록 과정](docs/pypi.md)
 - [버전 관리 방식](docs/versioning.md)
