+++ src/cli_onprem/__main__.py
@@ -7,7 +7,7 @@ from typing import Any, cast
 import typer
 from rich.console import Console
 
-# from cli_onprem.commands import docker_tar, fatpack, helm, s3_share
+# from cli_onprem.commands import docker_tar, tar_fat32, helm, s3_share
 
 context_settings = {
     "ignore_unknown_options": True,  # Always allow unknown options
@@ -33,7 +33,7 @@ def get_command(import_path: str) -> typer.Typer:
 
 
 app.add_typer(get_command("cli_onprem.commands.docker_tar:app"), name="docker-tar")
-app.add_typer(get_command("cli_onprem.commands.fatpack:app"), name="fatpack")
+app.add_typer(get_command("cli_onprem.commands.tar_fat32:app"), name="tar-fat32")
 app.add_typer(get_command("cli_onprem.commands.helm_local:app"), name="helm-local")
 app.add_typer(get_command("cli_onprem.commands.s3_share:app"), name="s3-share")
