+++ src/cli_onprem/commands/__init__.py
@@ -1,5 +1,5 @@
 """Command modules for CLI-ONPREM."""
 
-from . import docker_tar, fatpack, helm_local, s3_share
+from . import docker_tar, helm_local, s3_share, tar_fat32
 
-__all__ = ["docker_tar", "fatpack", "helm_local", "s3_share"]
+__all__ = ["docker_tar", "tar_fat32", "helm_local", "s3_share"]
